+++ src/widgets/search/__stories__/index.stories.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+import _ from "lodash";
+import { storiesOf } from "@storybook/react";
+
+import { Widget } from "components/widget";
+
+const Story = () => {
+  return (
+    <Widget
+      id="search-01"
+      type="search"
+      options={{ pattern: "https://duckduckgo.com/?q=%s", title: "DuckDuckGo" }}
+      data={{}}
+      setDataValue={_.noop}
+      setOptionValue={_.noop}
+      isLayoutEditable={false}
+      hasError={false}
+    />
+  );
+};
+
+storiesOf("Widgets.Search", module).add("Variants", () => <Story />);