+++ src/components/drawer/index.tsx
@@ -1,4 +1,5 @@
 import React, { memo } from "react";
+import { useTranslation } from "react-i18next";
 
 import Button, { ButtonSize } from "components/button";
 import { availableWidgetNames } from "widgets/index";
@@ -6,16 +7,21 @@ import Icon from "../icon";
 
 /* TODO: render display name instead of slug */
 const Drawer = memo(({ addWidget }: Props) => {
+  const { t } = useTranslation();
   return (
-    <div className="p-3 bg-color-default border" style={{ width: "200px" }}>
-      <div className="p-3 text-center uppercase font-bold text-2">Widgets</div>
+    <div className="p-3 bg-color-default border" style={{ width: "250px" }}>
+      <div className="p-3 text-center uppercase font-bold text-2">
+        {t("widget.common.drawerHeadline")}
+      </div>
       {availableWidgetNames.map(widgetName => (
         <div className="flex justify-between py-2">
-          {widgetName}
+          {t(`widget.${widgetName}.name`)}
           <Button
             size={ButtonSize.Small}
             outline
-            ariaLabel={`Add widget '${widgetName}'`}
+            ariaLabel={t("widget.common.add", {
+              widget: t(`widget.${widgetName}.name`)
+            })}
             onClick={() => addWidget(widgetName)}
           >
             <Icon name="plus" alt="" />