+++ src/components/language-select/selectors.ts
@@ -1,0 +1,10 @@
+import createSelector from "selectorator";
+
+const selectComponentProps = createSelector(
+  ["config.language"],
+  language => ({
+    language
+  })
+);
+
+export default selectComponentProps;