+++ src/components/loading/index.tsx
@@ -1,4 +1,5 @@
 import React, { memo } from "react";
+import { useTranslation } from "react-i18next";
 import cn from "classnames";
 
 import "./styles.scss";
@@ -7,8 +8,14 @@ export interface Props {
   className?: string;
 }
 
-const Loading = memo(({ className }: Props) => (
-  <div className={cn("loading mx-auto", className)} aria-label="Loading..." />
-));
+const Loading = memo(({ className }: Props) => {
+  const { t } = useTranslation();
+  return (
+    <div
+      className={cn("loading mx-auto", className)}
+      aria-label={t("common.loading")}
+    />
+  );
+});
 
 export default Loading;