+++ src/setupTests.js
@@ -2,4 +2,6 @@
 import { configure } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 
+import "common/i18n";
+
 configure({ adapter: new Adapter() });